Virtual_AI_Assitant/requirements.txt
speechrecognition
pyaudio
setuptools
pyttsx3
pocketsphinx
google.generativeai